[shell-ui] Commit C++ project structure. Add ASCII title.
This C++ project's structure is a complex one, consisting of a main
Makefile and separate so-called 'modules' with each one containing some
separate functionality of the whole project. See the main Makefile, as
well as each module's '.mk' files for additional information and for
configuration options.

Furthermore, this project utilizes some external C++ libraries namely:
- jsoncpp: presents functionality to parse JSON files;
- catch2: modern C++ unit testing framework.
Sidenote: as of this commit, neither of these libraries are utilized but
will be in a future update.
Also, this project is built with the help of the 'ncurses' library.

Work has begun on the client's terminal user interface, more
specifically - the title screen. The latter currently only displays the
game's ASCII title.