[shell-ui] Display title on main menu.
The title is made of ASCII symbols which form ten cards, each displaying
a separate letter of this project's name - "poke-shell".
This is implemented via the new 'Title' class. It essentially stores an
ncurses window and has methods which can manipulate the latter and print
the title string within said window.

Changelog:
- remove 'jsoncpp' from project's third party libraries. Its files are
still present within the project, but it is no longer compiled. The
reasoning being that it currently only add compilation time;
- add sub-namespaces to each source file of each module:
  - source code in 'include/shell_ui/title.hh' and 'src/shell_ui/title.cc'
  has been placed inside the namespace 'title';
  - source code in 'include/ncurses_helper/colors.hh' and
  'src/ncurses_helper/colors.cc' has been placed inside the namespace
  'colors';
  - from now on this pattern will continue to be applied to each
  functional part of a module.
- in 'ncurses_helper/colors': store color names in colors map as instances
of C++ strings and not as c-string pointers;
- remove macro 'TITLE_LETTER_COLOR_PAIR_NAME' and replace it with a
strongly typed static string variable inside the new Title class. Macros
could cause a lot of errors and confusion as they are not strongly typed
and may be redefined by other source files. Storing such values in
static variables ensures type-safety, as well as proper namespacing and access;
- 'shell_ui/main.cc': display the title, which is centered horizontally.