[shell-ui] Introduce test helpers.
Test helpers provide some useful functionality when testing. They are
meant to remove redundant code which doesn't describe the behavior of the
tested code, but reveals implementation details of unit tests.

Also, the 'test_helper' file provides routines which setup and tear down
the testing environment. This commit includes two such functions which
setup a testing environment for use with the 'ncurses' library.