[shell-ui] Change return type of 'std_helper::get_longest_string()'.
The function now returns a std::string copy and not a const reference to
a std::string as before. The reason behind this change: there is no
guarantee that the searched vector life ends after the returned longest
option. That is why, the result of this function has be independent of
the parsed vector.