[shell-ui] Add main menu scene.
The main menu scene is essentially the opening title screen. It also
contains a menu, which is non-functional at the moment. The menu is to
present options to the player to register, login, etc.

Changelog:
- Add default arguments to 'NCursesObject::mvwprint' method so that it
could be called dynamically without the need of a cast from an
NCursesObject pointer/reference to a specific child.
- Add alias for string-scene object map in Scene class.
- Introduce the ability for Makefiles to create subdirectories for
object files, following its corresponding source code's directory
structure.