[shell-ui] Remove inclusion of C++ source files during compilation.
They were previously included with the idea of splitting a template
class's interface and implementation from one another. However, I have
decided that it is best to keep a template class's definition only in one
header file because a template class is not a standalone C++ class, but
only a template to create classes.